General code cleanup (replace tabs by spaces, etc.) Finalised merge of PRs #10, #11, #12 and #13 Methods/functions added: - list_country_codes() - set_guestlogin_settings_base() - set_site_connectivity() - set_site_country() - set_site_guest_access() - set_site_locale() - set_site_mgmt() - set_site_name() - set_site_ntp() - set_site_snmp() Examples added: - ap_upgrade_firmware.php - site_provisioning_example - update_ac-iw_ports.php
