+++ README.md
@@ -37,6 +37,7 @@ The class currently supports the following functions/methods to get/post/put/del
 - list_alarms()
 - list_aps() (deprecated but still available as alias)
 - list_clients()
+- list_country_codes()
 - list_current_channels()
 - list_dashboard()
 - list_devices()
@@ -73,9 +74,18 @@ The class currently supports the following functions/methods to get/post/put/del
 - set_ap_radiosettings()
 - set_device_settings_base()
 - set_guestlogin_settings()
+- set_guestlogin_settings_base()
 - set_locate_ap() (deprecated but still available as alias)
 - set_networksettings_base()
 - set_radius_account_base()
+- set_site_connectivity()
+- set_site_country()
+- set_site_guest_access()
+- set_site_locale()
+- set_site_mgmt()
+- set_site_name()
+- set_site_ntp()
+- set_site_snmp()
 - set_sta_name()
 - set_sta_note()
 - set_usergroup()
