+++ examples/ap_upgrade_firmware.php
@@ -1,33 +1,37 @@
 <?php
-
-/** 
-* Checks and upgrades AP firmware (can be scheduled with systemd/cron)
-**/
-
+/**
+ * PHP API usage example
+ *
+ * contributed by: @4oo4
+ * description: example script to check and upgrade device firmware (can be scheduled with systemd/cron)
+ *              to the most current version
+ */
 require_once('vendor/autoload.php');
 require_once('config.php');
 
-// Because of a bug in the API, the site name is probably stuck at 'default' rather than what you actually named it:
-// https://github.com/Art-of-WiFi/UniFi-API-browser/issues/35
-$site_id = 'default';
-
-// AP MAC address formatted with colons
-$device_mac = 'de:ad:be:ef:01:23';
+/**
+ * site id of the AP to update
+ */
+$site_id = '<enter your site id here>';
 
 /**
- * initialize the Unifi API connection class, log in to the controller and request the alarms collection
- * (this example assumes you have already assigned the correct values to the variables used)
+ * device MAC address formatted with colons, e.g. 'de:ad:be:ef:01:23'
  */
+$device_mac = '<enter MAC address of device to update>';
 
+/**
+ * initialize the UniFi API connection class, log in to the controller
+ * (this example assumes you have already assigned the correct values in config.php to the variables used)
+ */
 $unifi_connection = new UniFi_API\Client($controlleruser, $controllerpassword, $controllerurl, $site_id, $controllerversion, false);
 $login            = $unifi_connection->login();
 
-// Run the actual upgrade
+/**
+ * Run the actual upgrade
+ */
 $results = $unifi_connection->upgrade_device($device_mac);
 
 /**
  * provide feedback in json format from $response given by upgrade_device();
  */
 echo json_encode($results, JSON_PRETTY_PRINT);
-
-?>
