+++ examples/extend_guest_auth.php
@@ -26,25 +26,25 @@ $set_debug_mode   = $unifi_connection->set_debug($debug);
 $loginresults     = $unifi_connection->login();
 
 if ($loginresults === 400) {
-	print "UniFi controller login failure, please check your credentials in config.php.\n";
+    print "UniFi controller login failure, please check your credentials in config.php.\n";
 } else {
-	$guestlist = $unifi_connection->list_guests();
-	// print "<pre>"; print_r ($guestlist); print "</pre>";
-	// loop thru all known guests
-	foreach ($guestlist as $guest) {
-		// print "<pre>"; print_r ($guest); print "</pre>";
-		print "<pre>" . $guest->_id . " (" . $guest->mac . "), valid until " . date (DATE_ATOM, $guest->end) . " (" . $guest->end . ")</pre>";
+    $guestlist = $unifi_connection->list_guests();
+    // print "<pre>"; print_r ($guestlist); print "</pre>";
+    // loop thru all known guests
+    foreach ($guestlist as $guest) {
+        // print "<pre>"; print_r ($guest); print "</pre>";
+        print "<pre>" . $guest->_id . " (" . $guest->mac . "), valid until " . date (DATE_ATOM, $guest->end) . " (" . $guest->end . ")</pre>";
 
-		// just a sample: only extend validity of guests which have end date after 2017-04-02
-		if ($guest->end > 1491166482) {
-			// extend clients five times = five days
-			if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
-			if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
-			if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
-			if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
-			if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
-		}
-	}
+        // just a sample: only extend validity of guests which have end date after 2017-04-02
+        if ($guest->end > 1491166482) {
+            // extend clients five times = five days
+            if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
+            if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
+            if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
+            if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
+            if (!$unifi_connection->extend_guest_validity ($guest->_id)) print "Extend failed for guest with id " . $guest->_id . "\n";
+        }
+    }
 
-	$logout_results = $unifi_connection->logout();
+    $logout_results = $unifi_connection->logout();
 }
